+++ src/main/java/net/mcbrincie/apel/lib/objects/ParticleCombiner.java
@@ -564,7 +564,12 @@ public class ParticleCombiner extends ParticleObject<ParticleCombiner> {
             ParticleObject<?> childObject = drawContext.getMetadata(OBJECT_IN_USE);
             // Defensive copy before passing to a child object
             Vector3f childDrawPos = new Vector3f(drawContext.getPosition());
-            childObject.doDraw(renderer, drawContext.getCurrentStep(), childDrawPos, drawContext.getNumberOfStep());
+            childObject.doDraw(renderer,
+                    drawContext.getCurrentStep(),
+                    childDrawPos,
+                    drawContext.getNumberOfStep(),
+                    drawContext.getDeltaTickTime()
+            );
             this.afterChildDraw.apply(drawContext, this);
         }
     }