+++ CHANGELOG.md
@@ -16,7 +16,7 @@ to use in their own projects
 - [ ] Added ``ParticleTrial`` for making a trial of particle objects
 - [x] Added ``roundness`` to ``ParticlePolygon`` for rounding the shape
 - [ ] Added ``scale`` to ParticleObject. Used for adjusting the scale (can also be negative)
-- [ ] Added ``timeDiff`` to DrawContext, which is used for knowing the difference of time between the steps
+- [x] Added ``deltaTimeTick`` to DrawContext, which is used for knowing the difference of time between the last tick and current tick
 - [x] Added ``numberOfSteps`` to DrawContext, which is used for knowing the number of steps of the path animator
 - [x] Support for 1.21.4